+++ src/OpenCqrs.EventSourcing.Store.Cosmos/ICosmosDataStore.cs
@@ -225,7 +225,7 @@ public interface ICosmosDataStore : IDisposable
     /// }
     /// var aggregate = result.Value;
     /// </example>
-    Task<Result<T>> UpdateAggregateDocument<T>(IStreamId streamId,
+    Task<Result<T?>> UpdateAggregateDocument<T>(IStreamId streamId,
         IAggregateId<T> aggregateId, AggregateDocument? aggregateDocument,
         CancellationToken cancellationToken = default) where T : IAggregateRoot, new();
 }