+++ docs/Release-Notes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## OpenCQRS 7.1.5
+_**Released ??/09/2025**_
+- Get aggregate with apply new events false returns now null if the aggregate doesn't exist
+- Update aggregate returns null if aggregate doesn't exist
+
 ## OpenCQRS 7.1.4
 _**Released 13/09/2025**_
 - Upgrade Nuget packages to latest versions